requirements.txt
sqlalchemy
psycopg2
requests
python-dotenv
werkzeug
